migrate build and dev to electron-vite
