+++ apps/code/package.json
@@ -10,14 +10,15 @@
   },
   "scripts": {
     "setup": "bash bin/setup",
-    "dev": "node scripts/dev.mjs",
-    "start": "node scripts/dev.mjs",
-    "start:debug": "ELECTRON_INSPECT=5858 node scripts/dev.mjs",
-    "package": "node scripts/build.mjs && electron-builder build --dir --config electron-builder.config.cjs",
-    "package:dev": "FORCE_DEV_MODE=1 SKIP_NOTARIZE=1 node scripts/build.mjs && electron-builder build --dir --config electron-builder.config.cjs",
-    "make": "node scripts/build.mjs && electron-builder build --config electron-builder.config.cjs",
+    "dev": "electron-vite dev --watch",
+    "start": "electron-vite dev --watch",
+    "start:debug": "electron-vite dev --watch --inspect=5858",
+    "compile": "electron-vite build",
+    "package": "electron-vite build && electron-builder build --dir --config electron-builder.config.cjs",
+    "package:dev": "FORCE_DEV_MODE=1 SKIP_NOTARIZE=1 electron-vite build && electron-builder build --dir --config electron-builder.config.cjs",
+    "make": "electron-vite build && electron-builder build --config electron-builder.config.cjs",
     "make:linux": "bash scripts/build-linux-docker.sh",
-    "publish": "node scripts/build.mjs && electron-builder build --publish always --config electron-builder.config.cjs",
+    "publish": "electron-vite build && electron-builder build --publish always --config electron-builder.config.cjs",
     "build": "pnpm package",
     "build-icons": "bash scripts/generate-icns.sh",
     "typecheck": "tsc -p tsconfig.node.json --noEmit && tsc -p tsconfig.web.json --noEmit",
@@ -60,6 +61,7 @@
     "adm-zip": "^0.5.16",
     "electron": "^41.0.0",
     "electron-builder": "^26.15.3",
+    "electron-vite": "^4.0.1",
     "husky": "^9.1.7",
     "jimp": "^1.6.0",
     "jsdom": "^26.0.0",
