+++ apps/code/scripts/postinstall.sh
@@ -12,7 +12,7 @@ SCRIPTS_DIR="$(cd "$(dirname "$0")" && pwd)"
 # pnpm skips package-level postinstall scripts when the lockfile is already
 # satisfied, so if node_modules/electron/dist gets wiped (interrupted download,
 # cache eviction, arch change, manual cleanup), `pnpm install` won't notice —
-# and `node scripts/dev.mjs` then fails with "Electron failed to install
+# and `electron-vite dev` then fails with "Electron failed to install
 # correctly, please delete node_modules/electron and try installing again".
 # Detect the missing binary and invoke Electron's own install script to fetch it.
 ELECTRON_DIST="$REPO_ROOT/node_modules/electron/dist"
