+++ knip.json
@@ -13,11 +13,9 @@
         "src/renderer/main.tsx",
         "src/renderer/desktop-services.ts",
         "src/renderer/desktop-contributions.ts",
-        "vite.main.config.mts",
-        "vite.preload.config.mts",
-        "vite.renderer.config.mts",
+        "electron.vite.config.ts",
+        "vite-main-plugins.mts",
         "vite.shared.mts",
-        "vite.workspace-server.config.mts",
         "scripts/*.{ts,mjs}"
       ],
       "project": ["src/**/*.{ts,tsx}", "scripts/**/*.{ts,mjs}"],
