+++ packages/workspace-server/src/services/agent/agent.ts
@@ -400,7 +400,7 @@ export class AgentService extends TypedEventEmitter<AgentServiceEvents> {
   }
 
   private getClaudeCliPath(): string {
-    // Keep in sync with the destDir in apps/code/vite.main.config.mts
+    // Keep in sync with the destDir in apps/code/vite-main-plugins.mts
     // (copyClaudeExecutable plugin).
     const binary = process.platform === "win32" ? "claude.exe" : "claude";
     return this.bundledResources.resolve(`.vite/build/claude-cli/${binary}`);
